Add files via upload
Caesar Cipher Encryption & Decryption Tool

The Caesar Cipher Encryption & Decryption Tool is a simple web-based project designed to demonstrate one of the earliest and most well-known encryption techniques. Developed using HTML, CSS, and JavaScript, this tool allows users to encrypt and decrypt messages by applying a specific numeric shift value to each letter in the text. When a message and a shift value are entered, the algorithm shifts each character by the specified number of positions in the alphabet, producing an encoded output